[Sb refactor] Remove some unused fields from MobileIconInteractor
Now that MobileIconInteractor provides an [icon] field, we can remove
some unused public fields that were all being calculated in the view
model. This simplifies the overall structure of the interactor -> view
model relationship, but it comes at the cost of requiring us to use the
real interactor implementations in tests.

Oh did I mention I updated the MobileIconViewModel tests to use the real
interactors. You now have to be pretty familiar with the data model
itself to understand what is happening in the tests. And there is more
work to be done with the fake repos, which still rely on the
externally-defined MobileMappings class.

Test: tests in statusbar/pipeline/*
Bug: 265342892
Change-Id: Ie85d92ad9d29f9d6d0c45ad72ca1c5e51482229b