Remove pinned tasks when starting locked task w/ intent
When a task is started w/ flags from an intent, it directly calls into
LockTaskController#startLockTaskMode. This misses the other code called
normally in ActivityTaskManagerService#startLockTaskMode such as closing
any pinned apps.

We fix this by moving this logic down a level to the common code path
shared by both

Bug: 295207389
Test: repro in bug
Change-Id: Id7386eae96cd014bf41ab3c4184ba9f3a79bdf10