Return valid task id of bottom activity for isTaskRoot
To match the documentation of Activity#isTaskRoot:
  The root is the first activity in a task.

Previously, the result may be affected by relinquishTaskIdentity.
For example:
Back from activity B to activity A[relinquishTaskIdentity=true]
in the same task. Activity#isTaskRoot will return true for B
because A uses the relinquish flag. That will cause
EnterTransitionCoordinator to skip the exit transition because
its mPendingExitNames will not be set.

Fix: 297013131
Test: atest TaskTests#testGetRootActivity_relinquishTaskIdentity
Change-Id: Iefa6e46c59a198f78a2f0495a20dc1239e04448c