Make WebViews eligible for Autofill For All Apps (AFAA)
AFAA, when enabled, ignores isImportantForAutofill flag on a View subclass. However, it only does so when view.getAutofillType() is something other than NONE. Autofill type could be a TEXT or DATE etc, so that EditTexts are of type TEXT, for example. A WebView is not of any type since it's just a system for hosting other components. Therefore this fix skips the autofillType check for a WebView.

Note that we need to always treat a WebView as important for autofill, otherwise all its contents (e.g. children views like EditTexts within the WebView) would be skipped entirely during the detection stage. (Note that for this reason WebView's isImportantForAutofill property is already `true` by default.)

Change-Id: Iade54fac30a6ff21ab58bf838d4e4845ebe8b686
Fix: 286195872
Test: atest CtsAssistTestCases