Launch notes task when the stylus tail button goes UP
The notes task will be shown immediately on the UP event of the stylus
tail button, instead of on the DOWN event, like before.

To do this, the framework must send all tail button events (both DOWN
and UP events) to SystemUI, because there is a CTS test that enforces
that SystemUI receives the DOWN event from a stylus. We do not have the
ability to modify the CTS test for this release.

Bug: 298056902
Test: manual with stylus
Test: atest NoteTaskInitializerTest
Change-Id: I12a08801fb767e495076051c049c11c9e635a58e