[SB][Flexi] Add status bar to lockscreen in flexiglass.
Importantly: Right now, NPVC passes a lot of different bits of
information to the keyguard status bar view. However in flexiglass, NPVC
does not exist. So, there's a lot of future work in having keyguard
status bar view fetch those bits of information without NPVC. This CL
is just a starting point :)

Notable bugs so far:
 - The status bar is visible on AOD when it should be hidden.
 - If you unplug then replug the device, the status bar on keyguard is
   no longer shown. (I think this is related to the battery charging
   animation.)

 Note on constraining the height:
 In flexiglass, KeyguardStatusBarView is detached & re-attached, so we
need to ensure the height is constrained in the constraint set so the
height is re-set each time it's re-attached. (In the old system, the
view was never detached, so the height didn't need to be re-set.)

Bug: 296122465
Test: enable flexiglass -> verify status bar with real icons appears on
lockscreen
Test: disable flexiglass -> verify status bar on lockscreen still works
Test: enable flexiglass. On lockscreen, open then close the shade ->
verify status bar is still visible after closing the shade
Test: atest KeyguardStatusBarViewControllerTest

Change-Id: Ie45f00834e9ff7fb8d2d3e8307ccb7818cac32e4