Ensure keyguard shows, and tells WM
Many race conditions can occur, especially with SideFPS. In this case,
a successful auth followed by numerous power button pushes, and having
distinct home and lockscreen wallpapers, produced a scenario where
keyguard was correctly thinking it should be visible but was not
always telling WM. This left the device in a state with just a blank
screen, behaving in a manner like it was occluded by another activity.

Always tell WM that sysui is showing, as there should be no harm in
telling it again (just in case).

Fixes: 301409430
Fixes: 300899918
Test: race condition - Enable 2 different wallpapers on
home/lockscreen, setup sidefps, authenticate, and slam the power
button over and over. This is easier to accomplish if you disabling
the double-tap camera launch gesture

Change-Id: I5c7137dfcf20fb27c697a60df21141d15979e03e