Only apply offset to the current window
Before this change, WallpaperController applies the wallpaper offsets to
all wallpaper window tokens. But when there are two wallapper windows
for home and lock screen, we don't want the lockscreen wallpaper to be
affected by the parallax on home screen from the launcher.

In order to do so, each wallpaper token will store the offsets, instead
of using global variables in WallpaperController. Update the dump
accordingly.

The zoom logic is unchanged (still a shared mLastWallpaperZoomOut for
all windows) but slightly refactored to not go through
updateWallpaperOffsetLocked.

Bug: 300544050
Test: atest WallpaperManagerTest
Test: atest WmTests:WallpaperControllerTests
Test: manually check that the bug is fixed
Test: manually go in the picker from first/third launcher page and make
sure all offsets for previews are correct
Test: "adb shell dumpsys activity" and check WallpaperController dump

Change-Id: I6287fd2e4144a8d5d82e463e440f189cc551ac8b