[flexiglass] Added isUserInputOngoing to transition model
Adds the concept of isUserInputOngoing to the scene transition model. This can be used to determine whether user input is driving a transition. Consumers of flinging state in the old world can use this value if they were relying on NPVC's definition of "flinging" which was actually just animating without user input.

Bug: 303321199
Test: added to SwipeToScene test
Change-Id: Ie25e1a6c8d64f32b854de36c99fbefa9fb5f3402