Fix context and resources used for keyguard presentation
This avoids creating a new context if the default context in the dagger graph is already the correct one for the display in the scope.

Before this cl, a new context was created also for the default display, and it had the wrong android.R.attr.isLightTheme value.

Ideally we should use the DisplaySpecific context in ClockEventController as well, but we first need to fix isLightTheme wrong value in the new display specific context.

Test: Manually check clocks colors + KeyguardClockSwitchScreenshotTest
Bug: 299256162
Change-Id: I942c8e35a644df27a6f442d7eddf8a2a767765be