Synchronize window config updates (7/n)
Allow ClientTransaction to take more than one ActivityLifecycleItem.

To keep the client side behavior unchanged, this CL refactors to allow
ClientTransaction to take lifecycle request in the same way as
non-lifecyle request, so that they are queued in order.

As a refactor, this should be no-op until replacing the server side
deprecated methods to use #addTransactionItem.

Bug: 260873529
Test: atest FrameworksCoreTests:TransactionExecutorTests
Change-Id: Ida33994bc78673a548c137be50a51daed4e873db